{feat} Added feature to support restarting game and recording score to files {fix} Issue of food appearing at border
Updated the controller class to receive a request to restart the game if required. The game class also updated to
support the same. In addition also the game class has been added an additional method to store the scores in a
text file. (can be used for a future enhancement to show scores)