feat: Added feature to include multi-threading
refactor: used bool for restartReq flag

Updated the game class to partly include a multi-threading component...
A thread will be spawned at the start which will wait on a cond-var till the score reaches 10 and
creates a slow speed zone of 10 seconds. After which it will restore the speed.
A mutex is required while accessing speed.