feat: Added feature to include wall / obstacle through a small level of inheritance
refactor: Updated snake class to make use of inheritance

Created a new Object parent class from which snake class can be inherited.
Object class contains some members that can be common to the snake and as well as the new Wall class.
Also involves a virtual function to check if an x & y position is part of the object ( wall or snake ).
Changed the colors to represent food and walls... Food is now green